requirements.txt
mcp
fastmcp
uvicorn
starlette
python-dotenv
simple-salesforce
langchain-openai
langgraph
httpx
requests
cryptography
pyjwt
click
langchain_mcp_adapters
langchain